Tests and fixes for breakpoints (#84)

+++ src/Spice86/Emulator/Memory/DmaChannel.cs
@@ -192,7 +192,7 @@ public sealed class DmaChannel {
             uint sourceOffset = (uint)this.Count + 1 - (uint)this.TransferBytesRemaining;
 
             int count = Math.Min(this.TransferChunkSize, this.TransferBytesRemaining);
-            byte[]? source = memory.GetData(memoryAddress + sourceOffset, count);
+            byte[]? source = memory.GetData(memoryAddress + sourceOffset, (uint)count);
 
             count = device.WriteBytes(source);
